[logging][refactor] Update logging utilities and notebooks
- Removed old logging modules: log_to_tables_utils.py, logger.py, task_logger.py, testLogger.py
- Added new logging utilities: pipeline_logger_utils.py, task_logger_utils.py, widget_utils.py
- Marked task_logger-superceded.py for reference
- Updated __init__.py in io_utils and logging_utils
- Adjusted bronze and silver notebooks to align with new logging structure
